Changelog: Extract findPullRequestId() into its own file
